Move styling and configuration code into own files
