+++ src/styling.py
@@ -1,0 +1,38 @@
+# ----------------------------------------------------------------------------
+# STYLING
+# ----------------------------------------------------------------------------
+
+TACC_IFRAME_SIZE = {
+    "max-width" : "1060px", "max-height" : "980px" # THESE ARE SET TO FIT IN THE 1080x1000 TACC iFRAME.  CAN BE REMOVED IF THOSE CONSTRAINTS COME OFF
+}
+
+CONTENT_STYLE = {
+    "padding": "2rem 1rem",
+    "font-family": 'Arial, Helvetica, sans-serif',
+}
+
+export_style = '''
+    position:absolute;
+    right:25px;
+    bottom:-55px;
+    font-family: Arial, Helvetica, sans-serif;
+    margin: 10px;
+    color: #fff;
+    background-color: #17a2b8;
+    border-color: #17a2b8;
+    display: inline-block;
+    font-weight: 400;
+    text-align: center;
+    white-space: nowrap;
+    vertical-align: middle;
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+    border: 1px solid transparent;
+    padding: .375rem .75rem;
+    font-size: 1rem;
+    line-height: 1.5;
+    border-radius: .25rem;
+    transition: color .15s ease-in-out,background-color .15s ease-in-out,border-color .15s ease-in-out,box-shadow .15s ease-in-out;
+'''
